+++ src/reporter/json_format.rs
@@ -143,6 +143,7 @@ mod tests {
                 since_commit: None,
                 branch: None,
             },
+            extra_ignore_comments: Vec::new(),
             content_filtering_args: ContentFilteringArgs {
                 max_file_size_mb: 25.0,
                 no_extract_archives: false,
@@ -163,6 +164,7 @@ mod tests {
             skip_regex: Vec::new(),
             skip_word: Vec::new(),
             no_base64: false,
+            no_inline_ignore: false,
         }
     }