+++ src/scanner/enumerate.rs
@@ -144,6 +144,8 @@ pub fn enumerate_filesystem_inputs(
         Some(&matcher_stats),
         enable_profiling,
         Some(shared_profiler),
+        &args.extra_ignore_comments,
+        args.no_inline_ignore,
     )?;
     let blob_processor_init_time = Mutex::new(t1.elapsed());
     let make_blob_processor = || -> BlobProcessor {