+++ tests/int_bitbucket.rs
@@ -132,6 +132,8 @@ fn test_bitbucket_remote_scan() -> Result<()> {
         skip_regex: Vec::new(),
         skip_word: Vec::new(),
         no_base64: false,
+        extra_ignore_comments: Vec::new(),
+        no_inline_ignore: false,
     };
 
     let global_args = GlobalArgs {